seperate logic for reading/exporting csv
