+++ src/test/java/alex220mark/utils/CsvReaderWriterTest.java
@@ -21,16 +21,4 @@ public class CsvReaderWriterTest {
 		assertLinesMatch(spaces, spacesRemoved);
 	}
 
-	@Test
-	void itShouldCapitalise() {
-		List<String> spaces = List.of("Red Socks", "Red Socks", "Red Socks", "Red Socks", "Red Socks", "Red Socks");
-		List<String> spacesRemoved = List.of(underTest.capitalizeString("red socks"),
-				underTest.capitalizeString("Red socks"),
-				underTest.capitalizeString("red Socks"),
-				underTest.capitalizeString("Red Socks"),
-				underTest.capitalizeString("reD soCKs"),
-				underTest.capitalizeString("rEd SOckS"));
-		assertLinesMatch(spaces, spacesRemoved);
-	}
-
 }
